Move function to file where it's used
